Implement a custom Pagefind UI
This was an interesting experience. Doing it this way does not feel
scalable. For one thing, there was a lot of back and forth between the
HTML and the JS file, as the JS is highly coupled to specifics of the
HTML.

I've designed it so that it should be easy to integrate it into the nav
bar if I want to, even having the results in a dropdown.

I'm also planning to add filtering. Lets get this out as is though.