Implement filtering of search results
This was a bit of a nightmare to write. It works, and I think it's about
as small and efficient as it is possible to be. But I never want to
write code like this again if I can help it. Declaritive UI libraries
like React really do have an advantage for this kind of work.